refactor(routes): extract /github into github.js (god-file decomposition)
Route contract unchanged (284 routes), 0 /github left inline, vitest 38 passed 268 passed . Lucy (GLM-5.2), driver-gated.
